check if we need to logout after managing account
