+++ apps/tlon-mobile/src/screens/ProfileScreen.tsx
@@ -1,57 +1,22 @@
 import { NativeStackScreenProps } from '@react-navigation/native-stack';
-import * as api from '@tloncorp/shared/dist/api';
 import * as store from '@tloncorp/shared/dist/store';
 import { ProfileScreenView, View } from '@tloncorp/ui';
-import * as Application from 'expo-application';
 import { useCallback } from 'react';
-import { Platform } from 'react-native';
 
-import { NOTIFY_PROVIDER, NOTIFY_SERVICE } from '../constants';
-import { clearShipInfo, useShip } from '../contexts/ship';
+import { useHandleLogout } from '../hooks/handleLogout';
 import { useCurrentUserId } from '../hooks/useCurrentUser';
-import { purgeDb } from '../lib/nativeDb';
 import NavBar from '../navigation/NavBarView';
 import { RootStackParamList } from '../types';
-import { removeHostingToken, removeHostingUserId } from '../utils/hosting';
 
 type Props = NativeStackScreenProps<RootStackParamList, 'Profile'>;
 
-const DEBUG_MESSAGE = `
-  Version: 
-  ${Platform.OS === 'ios' ? 'iOS' : 'Android'} ${Application.nativeBuildVersion}
-  
-  Notify Provider: 
-  ${NOTIFY_PROVIDER}
-
-  Notify Service: 
-  ${NOTIFY_SERVICE}
-`;
-
 export default function ProfileScreen(props: Props) {
-  const { clearShip } = useShip();
   const currentUserId = useCurrentUserId();
   const { data: contacts } = store.useContacts();
-
-  const handleLogout = useCallback(async () => {
-    await purgeDb();
-    api.queryClient.clear();
-    api.removeUrbitClient();
-    clearShip();
-    clearShipInfo();
-    removeHostingToken();
-    removeHostingUserId();
-  }, [clearShip]);
-
-  const onManageAccountPressed = useCallback(() => {
-    props.navigation.navigate('ManageAccount');
-  }, [props.navigation]);
+  const handleLogout = useHandleLogout();
 
   const onAppSettingsPressed = useCallback(() => {
     props.navigation.navigate('AppSettings');
-  }, [props.navigation]);
-
-  const onBlockedUsersPressed = useCallback(() => {
-    props.navigation.navigate('BlockedUsers');
   }, [props.navigation]);
 
   return (
@@ -59,11 +24,8 @@ export default function ProfileScreen(props: Props) {
       <ProfileScreenView
         contacts={contacts ?? []}
         currentUserId={currentUserId}
-        debugMessage={DEBUG_MESSAGE}
-        onManageAccountPressed={onManageAccountPressed}
         onAppSettingsPressed={onAppSettingsPressed}
-        onBlockedUsersPressed={onBlockedUsersPressed}
-        handleLogout={handleLogout}
+        onLogoutPressed={handleLogout}
       />
       <NavBar navigation={props.navigation} />
     </View>
